Move building (sidebar) and map components to nested dirs
