+++ app/src/frontend/building/fields-config.json
