+++ app/src/frontend/building/multi-edit.tsx
@@ -5,8 +5,8 @@ import PropTypes from 'prop-types';
 
 import Sidebar from './sidebar';
 import CONFIG from './fields-config.json';
-import InfoBox from './components/info-box';
-import { sanitiseURL } from './helpers';
+import InfoBox from '../components/info-box';
+import { sanitiseURL } from '../helpers';
 
 const MultiEdit = (props) => {
     if (!props.user){
