+++ app/src/frontend/map/search-box.css
