+++ app/src/frontend/map/theme-switcher.tsx
